feat: 🎸 Authentication
This feature provides a way for authenticating using a filter for each
request. First, the application creates a user with a criptograpted
username and password (view `DemoApplication.java`). The requests are
filtered by a `UsernamePasswordAuthenticationFilter`, whitch is a
default way to authenticate a user. This authentication is provided
throw the userDetailService configured at `CustomUserDetailsService`.
You can test it by sending a post request to
http://localhost:8080/login?password=password&username=username